refactor(dh): move component to folder
